remove GHH_LIB_INTERNAL defs, move timeit and gtimer to time.c
