+++ src/array.c
@@ -1,4 +1,3 @@
-#define GHH_LIB_INTERNAL
 #include <stdlib.h>
 #include <stdbool.h>
 #include <ghh/array.h>
